pyproject.toml
[tool.poetry]
name = "avito-banners"
version = "0.1.0"
description = ""
authors = ["daniskzn <danis.sharafiev@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pyjwt = "^2.8.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
pydantic-settings = "^2.2.1"
ruff = "^0.3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
